catalog/requirements.txt
dj-database-url==0.5.0
Django==3.1.2
djangogrpcframework==0.2.1
djangorestframework==3.12.4
grpcio-tools~=1.30
gunicorn==20.0.4
psycopg2-binary==2.8.6
qservices-library==1.2
wheel==0.35.1
whitenoise==5.2.0
